+++ lib/src/validator/allValidations.dart
@@ -1,3 +1,0 @@
-// Deprecated: use all_validations.dart
-// ignore: unused_import
-export 'all_validations.dart';